requirements.txt
google-genai
openai
pandas
sentence-transformers
scikit-learn
datasets
fastapi
uvicorn
python-dotenv
